Added src/train.py
